1.making the gallery model for the images and album including all the necessary attrubute 2.create page foe the album model 3.multiple image entry possible 4.on click on the icon the multiple image in the index pagge is showing 5.error in the user section solved 6.new pagination instead of the default one 7.make some changes in the gallery image section .

+++ app/Http/Controllers/BlogCategoryController.php
@@ -196,51 +196,53 @@ class BlogCategoryController extends Controller
         if ($existimage != $currentimage) {
 
             // Delete old images if they exist
-            if ($blogCategory->image) {
-                if (Storage::exists(public_path('storage/'.$blogCategory->image))) {
-                    Storage::delete(public_path('storage/'.$blogCategory->image));
+            if ($request->input('image')) {
+                if ($blogCategory->image) {
+                    if (Storage::exists(public_path('storage/'.$blogCategory->image))) {
+                        Storage::delete(public_path('storage/'.$blogCategory->image));
+                    }
+                    if (Storage::exists(public_path('storage/images/resized/800px_'.basename($blogCategory->image)))) {
+                        Storage::delete(public_path('storage/images/resized/800px_'.basename($blogCategory->image)));
+                    }
+                    if (Storage::exists(public_path('storage/images/resized/100px_'.basename($blogCategory->image)))) {
+                        Storage::delete(public_path('storage/images/resized/100px_'.basename($blogCategory->image)));
+                    }
                 }
-                if (Storage::exists(public_path('storage/images/resized/800px_'.basename($blogCategory->image)))) {
-                    Storage::delete(public_path('storage/images/resized/800px_'.basename($blogCategory->image)));
-                }
-                if (Storage::exists(public_path('storage/images/resized/100px_'.basename($blogCategory->image)))) {
-                    Storage::delete(public_path('storage/images/resized/100px_'.basename($blogCategory->image)));
-                }
+
+                $imagePath = $request->input('image');
+                $filename = basename($imagePath);
+
+                //path for the original image and resized image
+                $originalPath = 'images/'.$filename;
+                $resized100Path = 'images/resized/100px_'.$filename;
+                $resized800Path = 'images/resized/800px_'.$filename;
+                // $resizedPath = 'images/resized/'.$filename;
+
+                //move the file from temporary to original place
+                Storage::disk('public')->move($imagePath, $originalPath);
+
+                //resize of 100px
+                $resized100Image = Image::make(storage_path('app/public/'.$originalPath))->resize(100, null, function ($constraint) {
+                    $constraint->aspectRatio(); // Keep aspect ratio
+                    $constraint->upsize(); // Prevent upsizing
+                });
+
+                //store the image in the resized folder
+                Storage::disk('public')->put($resized100Path, (string) $resized100Image->encode());
+
+                // for the resized 800
+                $resized800Image = Image::make(storage_path('app/public/'.$originalPath))->resize(800, null, function ($constraint) {
+                    $constraint->aspectRatio(); // Keep aspect ratio
+                    $constraint->upsize(); // Prevent upsizing
+                });
+                Storage::disk('public')->put($resized800Path, (string) $resized800Image->encode());
+
+                //saved in the database
+                $blogCategory->image = $originalPath;
+            } else {
+
+                $blogCategory->image = $blogCategory->image;
             }
-
-            $imagePath = $request->input('image');
-            $filename = basename($imagePath);
-
-            //path for the original image and resized image
-            $originalPath = 'images/'.$filename;
-            $resized100Path = 'images/resized/100px_'.$filename;
-            $resized800Path = 'images/resized/800px_'.$filename;
-            // $resizedPath = 'images/resized/'.$filename;
-
-            //move the file from temporary to original place
-            Storage::disk('public')->move($imagePath, $originalPath);
-
-            //resize of 100px
-            $resized100Image = Image::make(storage_path('app/public/'.$originalPath))->resize(100, null, function ($constraint) {
-                $constraint->aspectRatio(); // Keep aspect ratio
-                $constraint->upsize(); // Prevent upsizing
-            });
-
-            //store the image in the resized folder
-            Storage::disk('public')->put($resized100Path, (string) $resized100Image->encode());
-
-            // for the resized 800
-            $resized800Image = Image::make(storage_path('app/public/'.$originalPath))->resize(800, null, function ($constraint) {
-                $constraint->aspectRatio(); // Keep aspect ratio
-                $constraint->upsize(); // Prevent upsizing
-            });
-            Storage::disk('public')->put($resized800Path, (string) $resized800Image->encode());
-
-            //saved in the database
-            $blogCategory->image = $originalPath;
-        } else {
-
-            $blogCategory->image = $blogCategory->image;
         }
         $blogCategory->save();
